No scroll view No mini-map
