+++ Assets/Logger.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: c4d82a0c6142a6b4eb0f2143d6c9cc15
+guid: 2d6d958d5f35cd84999de8e85a3938e4
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
